Diálogos
- Guna
- Bodia
- Velha
- Dealer

+++ Assets/Scripts/PlayerStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "newPlayerStats", menuName = "Scriptable Objects/Player Stats")]
 public class PlayerStats : ScriptableObject
 {
-    public float moneyAmount;
+    public int moneyAmount;
     public int drugsAmount;
     public bool hasWeapon;
     public int wantedLevel;
